+++ mr/src/itest/java/org/opensearch/hadoop/integration/rest/AbstractRestSaveTest.java
@@ -34,6 +34,7 @@ import java.util.Scanner;
 
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
+import org.junit.Ignore;
 import org.opensearch.hadoop.OpenSearchHadoopIllegalArgumentException;
 import org.opensearch.hadoop.cfg.ConfigurationOptions;
 import org.opensearch.hadoop.cfg.Settings;
@@ -66,9 +67,11 @@ public class AbstractRestSaveTest {
 
     private static final Log LOG = LogFactory.getLog(AbstractRestSaveTest.class);
 
+    private static final JsonUtils.Query HITS_TOTAL_VALUE = JsonUtils.query("hits").get("total").get("value");
+
     @Test
     public void testBulkWrite() throws Exception {
-        TestSettings testSettings = new TestSettings("rest/savebulk");
+        TestSettings testSettings = new TestSettings("rest_save_bulk");
         //testSettings.setPort(9200)
         testSettings.setProperty(ConfigurationOptions.OPENSEARCH_SERIALIZATION_WRITER_VALUE_CLASS, JdkValueWriter.class.getName());
         RestRepository client = new RestRepository(testSettings);
@@ -83,6 +86,7 @@ public class AbstractRestSaveTest {
             line.put("name", in.next());
             line.put("url", in.next());
             line.put("picture", in.next());
+            in.nextLine();
             client.writeToIndex(line);
             line.clear();
         }
@@ -90,9 +94,10 @@ public class AbstractRestSaveTest {
         client.close();
     }
 
+    @Ignore("OpenSearch throws an error on empty bulk request")
     @Test
     public void testEmptyBulkWrite() throws Exception {
-        TestSettings testSettings = new TestSettings("rest/emptybulk");
+        TestSettings testSettings = new TestSettings("rest_empty_bulk");
         testSettings.setInternalClusterInfo(TestUtils.getOpenSearchClusterInfo());
         testSettings.setProperty(ConfigurationOptions.OPENSEARCH_SERIALIZATION_WRITER_VALUE_CLASS, JdkValueWriter.class.getName());
         RestRepository restRepo = new RestRepository(testSettings);
@@ -105,8 +110,9 @@ public class AbstractRestSaveTest {
 
     @Test
     public void testRepositoryDelete() throws Exception {
-        Settings settings = new TestSettings("rest/deletebulk");
-        RestUtils.delete("rest");
+        String index = "rest_delete_bulk";
+        Settings settings = new TestSettings(index);
+        RestUtils.delete(index);
         InitializationUtils.discoverClusterInfo(settings, LOG);
         settings.setProperty(ConfigurationOptions.OPENSEARCH_SERIALIZATION_WRITER_VALUE_CLASS, JdkValueWriter.class.getName());
         settings.setProperty(ConfigurationOptions.OPENSEARCH_MAPPING_DEFAULT_EXTRACTOR_CLASS, ConstantFieldExtractor.class.getName());
@@ -120,18 +126,18 @@ public class AbstractRestSaveTest {
         String doc = "{\"index\":{\"_id\":\"" + StringUtils.jsonEncoding(id) + "\"}}\n{\"field\":1}\n";
         repository.writeProcessedToIndex(new BytesArray(doc));
         repository.flush();
-        RestUtils.refresh("rest");
+        RestUtils.refresh(index);
 
-        assertThat(JsonUtils.query("hits").get("total").apply(JsonUtils.asMap(RestUtils.get("rest/deletebulk/_search"))), is(equalTo(1)));
+        assertThat(HITS_TOTAL_VALUE.apply(JsonUtils.asMap(RestUtils.get(index + "/_search"))), is(equalTo(1)));
 
         repository.delete();
 
-        assertThat(JsonUtils.query("hits").get("total").apply(JsonUtils.asMap(RestUtils.get("rest/deletebulk/_search"))), is(equalTo(0)));
+        assertThat(HITS_TOTAL_VALUE.apply(JsonUtils.asMap(RestUtils.get(index + "/_search"))), is(equalTo(0)));
     }
 
     @Test
     public void testRepositoryDeleteEmptyIndex() throws Exception {
-        Settings settings = new TestSettings("delete_empty/test");
+        Settings settings = new TestSettings("delete_empty");
         RestUtils.delete("delete_empty");
         InitializationUtils.discoverClusterInfo(settings, LOG);
         settings.setProperty(ConfigurationOptions.OPENSEARCH_SERIALIZATION_WRITER_VALUE_CLASS, JdkValueWriter.class.getName());
@@ -143,11 +149,11 @@ public class AbstractRestSaveTest {
         RestRepository repository = new RestRepository(settings);
         repository.touch();
 
-        assertThat(JsonUtils.query("hits").get("total").apply(JsonUtils.asMap(RestUtils.get("delete_empty/test/_search"))), is(equalTo(0)));
+        assertThat(HITS_TOTAL_VALUE.apply(JsonUtils.asMap(RestUtils.get("delete_empty/_search"))), is(equalTo(0)));
 
         repository.delete();
 
-        assertThat(JsonUtils.query("hits").get("total").apply(JsonUtils.asMap(RestUtils.get("delete_empty/test/_search"))), is(equalTo(0)));
+        assertThat(HITS_TOTAL_VALUE.apply(JsonUtils.asMap(RestUtils.get("delete_empty/_search"))), is(equalTo(0)));
     }